requirements.txt
Django
djangorestframework
django-filter
markdown
Scrapy
apprise
celery
django-celery-beat
mmh3
Redis
lxml
beautifulsoup4